requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.3.0
MarkupSafe==3.0.3
numpy==2.4.1
psycopg2-binary==2.9.11
shapely==2.1.2
Werkzeug==3.1.5
